server/requirements.txt
fastapi==0.110.1
requests==2.31.0
uvicorn
pymongo[srv]
python-dotenv
openai==1.16.2
langchain==0.1.14
langchain_nomic==0.0.2